# darapi/bin/Release/net6.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "My API",
    "version": "v1"
  },
  "paths": {
    "/nasa_/post_daily_photo": {
      "post": {
        "tags": [
          "nasa_"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "chatid",
            "type": "integer",
            "format": "int64"
          },
          {
            "in": "query",
            "name": "datetime",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "string"
            }
          }
        }
      }
    },
    "/nasa_/get_asteroids": {
      "get": {
        "tags": [
          "nasa_"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "start_datetime",
            "type": "string"
          },
          {
            "in": "query",
            "name": "end_datetime",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Asteroid"
              }
            }
          }
        }
      }
    },
    "/nasa_/post_asteroids_list": {
      "post": {
        "tags": [
          "nasa_"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "id",
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/nasa_/put_asteroid_to_list": {
      "put": {
        "tags": [
          "nasa_"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "id",
            "type": "integer",
            "format": "int64"
          },
          {
            "in": "query",
            "name": "number",
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/Asteroid"
            }
          }
        }
      }
    },
    "/nasa_/post_my_asteroids_list": {
      "post": {
        "tags": [
          "nasa_"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "id",
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/nasa_/delete_asteroid_from_list": {
      "delete": {
        "tags": [
          "nasa_"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "id",
            "type": "integer",
            "format": "int64"
          },
          {
            "in": "query",
            "name": "number",
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/nasa_/user_is_subscribed_to_updates/{id}": {
      "put": {
        "tags": [
          "nasa_"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int64"
          },
          {
            "in": "query",
            "name": "b",
            "type": "boolean"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "string"
            }
          }
        }
      },
      "get": {
        "tags": [
          "nasa_"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "boolean"
            }
          }
        }
      }
    },
    "/nasa_/bot_is_waiting_for_number_of_asteroid_to_add/{id}": {
      "put": {
        "tags": [
          "nasa_"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int64"
          },
          {
            "in": "query",
            "name": "b",
            "type": "boolean"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "string"
            }
          }
        }
      },
      "get": {
        "tags": [
          "nasa_"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "boolean"
            }
          }
        }
      }
    },
    "/nasa_/bot_is_waiting_for_number_of_asteroid_to_delete/{id}": {
      "put": {
        "tags": [
          "nasa_"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int64"
          },
          {
            "in": "query",
            "name": "b",
            "type": "boolean"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "string"
            }
          }
        }
      },
      "get": {
        "tags": [
          "nasa_"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "boolean"
            }
          }
        }
      }
    },
    "/nasa_/bot_is_waiting_for_photo_data/{id}": {
      "put": {
        "tags": [
          "nasa_"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int64"
          },
          {
            "in": "query",
            "name": "b",
            "type": "boolean"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "string"
            }
          }
        }
      },
      "get": {
        "tags": [
          "nasa_"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "boolean"
            }
          }
        }
      }
    }
  },
  "definitions": {
    "Asteroid": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "epochDateCloseApproach": {
          "format": "int64",
          "type": "integer"
        },
        "kilometersPerSecond": {
          "format": "double",
          "type": "number"
        },
        "isPotentiallyHazardous": {
          "type": "boolean"
        },
        "estimatedDiameterMin": {
          "format": "double",
          "type": "number"
        },
        "estimatedDiameterMax": {
          "format": "double",
          "type": "number"
        }
      }
    }
  }
}